Convert BeaconPreset to type and rename file
